+++ src/types/frontier.h
@@ -11,7 +11,7 @@ struct Segment_t {
     char gs_visited;
     int gs_distance_to_initial;
     int gs_distance_to_target;
-    int gs_factor;
+    unsigned long long int gs_factor;
     
     int frontiers_count;
 };
